add new index flow

+++ everdues/contracts/root/EverduesRoot.sol
@@ -104,9 +104,9 @@ contract EverduesRoot is EverduesRootBase {
 				uint8,
 				mapping(address => ServiceDeployParams),
 				address,
-				mapping(uint8 => mapping(uint256 => ExternalSubscription)),
-				mapping(uint8 => string),
-				mapping (uint8=>string[]),
+				mapping(uint32 => mapping(uint256 => ExternalSubscription)),
+				mapping(uint32 => string),
+				mapping (uint32=>string[]),
 				address// TODO: add ,address (watcher)
 			)
 		);
